mysite/requirements.txt
django
djangorestframework
markdown
django-filter
celery
flower
redis